installスクリプトの改善

+++ install.sh
@@ -6,23 +6,41 @@ cp $OLD_PWD/diff-highlight /usr/local/bin/diff-highlight
 chmod +x /usr/local/bin/diff-highlight
 
 cd $HOME
-rm ~/.bash_profile
+if [ -e ~/.bash_profile ]; then
+    rm ~/.bash_profile
+fi
 ln -s git/dotfiles/.bash_profile
-rm ~/.bashrc
+if [ -e ~/.bashrc ]; then
+    rm ~/.bashrc
+fi
 ln -s git/dotfiles/.bashrc
-rm ~/.gitconfig
+if [ -e ~/.gitconfig ]; then
+    rm ~/.gitconfig
+fi
 ln -s git/dotfiles/.gitconfig
-rm ~/.git-completion.bash
+if [ -e ~/.git-completion.bash ]; then
+    rm ~/.git-completion.bash
+fi
 ln -s git/dotfiles/.git-completion.bash
-rm ~/.git-prompt.sh
+if [ -e ~/.git-prompt.sh ]; then
+    rm ~/.git-prompt.sh
+fi
 ln -s git/dotfiles/.git-prompt.sh
-rm ~/.hg-completion.bash
+if [ -e ~/.hg-completion.bash ]; then
+    rm ~/.hg-completion.bash
+fi
 ln -s git/dotfiles/.hg-completion.bash
-rm ~/.inputrc
+if [ -e ~/.inputrc ]; then
+    rm ~/.inputrc
+fi
 ln -s git/dotfiles/.inputrc
-rm ~/.screenrc
+if [ -e ~/.screenrc ]; then
+    rm ~/.screenrc
+fi
 ln -s git/dotfiles/.screenrc
-rm ~/.vimrc
+if [ -e ~/.vimrc ]; then
+    rm ~/.vimrc
+fi
 ln -s git/dotfiles/.vimrc
 
 touch .bashrc.local
@@ -40,7 +58,9 @@ fi
 cp $OLD_PWD/php.dict $HOME/.vim/dictionary/php.dict
 
 cd $HOME/.vim/
-rm phpcs-rule.xml
+if [ -e phpcs-rule.xml ]; then
+	rm phpcs-rule.xml
+fi
 ln -s ../git/dotfiles/phpcs-rule.xml
 if [ ! -e $HOME/.vim/plugin/bufexplorer.vim ]; then
 	curl 'https://codeload.github.com/jlanzarotta/bufexplorer/zip/master' -o bufexplorer.zip
